samplerun

+++ packer-local-test.sh
@@ -14,6 +14,17 @@ if [ $# -eq 0 ] || [[ "$@" =~ "-h" ]];then
   exit 0
 fi
 
+#check for go and gox binary
+#if no present, we skip local build, local test.
+unset CHECK
+which gox go bzr &>/dev/null
+if [ $? -eq 0 ]; then
+  CHECK=0
+else
+  echo "Info: bzr, go or gox not found, skipping local build"
+  CHECK=1
+fi
+
 GHTEST="`curl -sS -I https://api.github.com 2>&1`"
 RET=$?
 
@@ -24,33 +35,47 @@ if [ ${RET} -ne 0 ]; then
   exit ${RET}
 fi
 
-if [ -d packer ];then
-  pushd packer
+if [ $GOPATH ] && [ -d $GOPATH/src/github.com/mitchellh/packer ];then
+  pushd $GOPATH/src/github.com/mitchellh/packer
+  #check .git/config for pr/*, add if not present
+  grep 'fetch.*.=.*.+refs/pull/\*/head:refs/remotes/origin/pr/\*' .git/config &>/dev/null
+  if [ $? -ne 0 ]; then
+    git config --add remote.origin.fetch '+refs/pull/*/head:refs/remotes/origin/pr/*'
+  fi
   git clean -fdx
   git checkout master
   git pull -f --prune origin HEAD
   git fetch --all --tags
   popd
-else
-  git clone https://github.com/mitchellh/packer.git
-  pushd packer
+elif  [ $GOPATH ];then
+  go get github.com/mitchellh/packer
+  pushd $GOPATH/src/github.com/mitchellh/packer
   git config --add remote.origin.fetch '+refs/pull/*/head:refs/remotes/origin/pr/*'
   git checkout master
   git pull -f --prune origin HEAD
   git fetch --all --tags
+  #on first run we run make updatedeps
+  make updatedeps
   popd
 fi
 
-if [ "${1}" == "master" ] || [ "${1}" -eq 0 ] ;then
-  CHECKOUT="master"
-  echo "Will checkout to master"
-  pushd packer
-  git checkout master
-  popd
-elif [ "${1}" ];then
-  CHECKOUT="pr/${1}"
-  echo "Will checkout to ${CHECKOUT}"
-  pushd packer
-  git checkout ${CHECKOUT}
-  popd
+if [ $CHECK ] && [ $CHECK -eq 0 ]; then
+  echo "build goes here"
+  pushd $GOPATH/src/github.com/mitchellh/packer
+  #run make updatedeps
+  if [[ "${@}" =~ "updatedeps" ]] ;then
+    echo "running make updatedeps"
+    make updatedeps
+  fi
+  if [ "${1}" == "master" ] || [ "${1}" -eq 0 ] ;then
+    CHECKOUT="master"
+    echo "Will checkout to master"
+    git checkout ${CHECKOUT} --force
+  elif [ "${1}" ];then
+    CHECKOUT="pr/${1}"
+    echo "Will checkout to ${CHECKOUT}"
+    git checkout ${CHECKOUT} --force
+  fi
+  make test
+  make dev
 fi
